requirements.txt
argparse
requests
python-nmap
python-dotenv
beautifulsoup4
